mmr-testing/MMR-Source-url-set-to-childNode.html

<!DOCTYPE html>
<html>
<head>
    <title>Source is set to ChildNode test case.</title>
    <meta name="timeout" content="normal">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="mediasource-util.js"></script>
</head>
<body>
    <div id="log"></div>
    <video autoplay id="Video1" style="visibility: visible;">
        <source src="mp4/test.mp4" type="video/mp4; codecs=&quot;mp4a.40.2, avc1.4d400d&quot;">
    </video>
    <script>
        mediasource_test(function (test) {
            var mediaElement = document.getElementById("Video1");
            mediaElement.addEventListener('error', test.unreached_func("Unexpected event 'error'"));
            mediaElement.addEventListener('ended', test.step_func_done());

            test.expectEvent(mediaElement, 'playing', 'Playing triggered');

            test.step_timeout(function () {
                assert_false(test.eventExpectations_.expectingEvents(), "Pending event expectations.");
            }, 5000);

        }, "Test soure is set to childNode");

    </script>

</body>
</html>
